Better handling of cyclic reference and top level reference
*Cyclic reference*

This commits changes the behavior of the ucfg library when we detect
cyclic reference, before when a cyclic was defined in a configuration
the library would go into an infinite loop and panic because of the
stack.

The new behavior is the following, when a cyclic reference is detected
when resolving the reference it will try to resolve it with the other
configured resolvers (the environment and the keystore for beats). If the key
is not found in theses resolvers the error will be send back to the
users with the problematic key.

If the values is found in the other resolver it will return that value
and it wont be a cyclic reference.

The detection is using a `VisitedFieldsRecorder` that will keep track of
the paths that was required to generate the concrete value, if the path is see
twice we return an error.

The caching behavior of the types was changed to only take care of primitives
values an not cfgDynamic.

*Top level reference*

Change of behavior in the `resolveRef`, it will now return ErrMissing
when a top level key is not found so the key can be check with the other
resolvers.