Fix an issue when a cyclic reference was incorrectly detected when multiple reference was used.
Fix was to make sure when we call `doReifyPrimitives` we need to clear
fields before tring to make the value concrete, because this will
require to revisited some fields to generate the actual values.

Problematic structure
```
{
  "path": {
    "home": "/var",
    "logs" "${path.home}"
  }
  "output" : {
    "file": "${path.logs}"
  }
}
```

We could also improve the caching mechanism to solve that issues, but
this seems to be the simplest fix to do.

Fixes: https://github.com/elastic/beats/issues/6214